fixed build warning
